Move unwrap_graphql_error to utils
