+++ ariadne/logger.py
@@ -1,0 +1,8 @@
+import logging
+
+
+logger = logging.getLogger("ariadne")
+
+
+def log_graphql_error(error):
+    pass
