+++ ariadne/utils.py
@@ -1,3 +1,5 @@
+from typing import Optional, Union
+
 from graphql import parse
 
 
@@ -13,3 +15,11 @@ def convert_camel_case_to_snake(graphql_name: str) -> str:
 def gql(value: str) -> str:
     parse(value)
     return value
+
+
+def unwrap_graphql_error(
+    error: Union[GraphQLError, Optional[Exception]]
+) -> Optional[Exception]:
+    if isinstance(error, GraphQLError):
+        return unwrap_graphql_error(error.original_error)
+    return error
